Create 3D_PDB_NAG
Code takes in a PDB ID, downloads it from RCSB and outputs 2 files. The first file is a csv file containing all atoms that are part of the desired structure, the file is formatted as [chain][residue_number][residue_name][atom_name][x][y][z]. The second file is also in csv format, that contains all sites of Glycosylation and the plausible AA positions that are associated with them. These occasionally need to be checked manually. Overall, this code can be used as a prelude to a distance matrix calculation pipeline for all atoms in a structure.